+++ cps/src/test/java/org/jenkinsci/plugins/workflow/cps/SnippetizerTest.java
@@ -46,10 +46,13 @@ import org.jenkinsci.plugins.workflow.support.steps.ExecutorStep;
 import org.jenkinsci.plugins.workflow.support.steps.StageStep;
 import org.jenkinsci.plugins.workflow.support.steps.WorkspaceStep;
 import org.jenkinsci.plugins.workflow.support.steps.build.BuildTriggerStep;
+import org.jenkinsci.plugins.workflow.support.steps.input.InputStep;
 import org.junit.Test;
 import static org.junit.Assert.*;
 import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.jvnet.hudson.test.Email;
+import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.JenkinsRule;
 
 public class SnippetizerTest {
@@ -72,7 +75,12 @@ public class SnippetizerTest {
     @Test public void coreStep() throws Exception {
         ArtifactArchiver aa = new ArtifactArchiver("x.jar");
         aa.setAllowEmptyArchive(true);
-        assertRoundTrip(new CoreStep(aa), "step([$class: 'ArtifactArchiver', allowEmptyArchive: true, artifacts: 'x.jar', defaultExcludes: true, excludes: '', fingerprint: false, onlyIfSuccessful: false])");
+        assertRoundTrip(new CoreStep(aa), "step([$class: 'ArtifactArchiver', allowEmptyArchive: true, artifacts: 'x.jar'])");
+    }
+
+    @Ignore("TODO until 1.601+ expected:<step[([$class: 'ArtifactArchiver', artifacts: 'x.jar'])]> but was:<step[ <object of type hudson.tasks.ArtifactArchiver>]>")
+    @Test public void coreStep2() throws Exception {
+        assertRoundTrip(new CoreStep(new ArtifactArchiver("x.jar")), "step([$class: 'ArtifactArchiver', artifacts: 'x.jar'])");
     }
 
     @Test public void blockSteps() throws Exception {
@@ -98,6 +106,11 @@ public class SnippetizerTest {
                 (which does not matter in this case since BuildTriggerStep/config.jelly does not offer to bind parameters anyway)
         assertRoundTrip(step, "build job: 'downstream', parameters: [[$class: 'hudson.model.StringParameterValue', name: 'branch', value: 'default'], [$class: 'hudson.model.BooleanParameterValue', name: 'correct', value: true]]");
         */
+    }
+
+    @Issue("JENKINS-25779")
+    @Test public void defaultValues() throws Exception {
+        assertRoundTrip(new InputStep("Ready?"), "input 'Ready?'");
     }
 
     private static void assertRoundTrip(Step step, String expected) throws Exception {