+++ support/src/test/java/org/jenkinsci/plugins/workflow/support/steps/input/InputStepTest.java
@@ -1,0 +1,58 @@
+/*
+ * The MIT License
+ *
+ * Copyright 2015 Jesse Glick.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package org.jenkinsci.plugins.workflow.support.steps.input;
+
+import java.util.Collections;
+import org.jenkinsci.plugins.workflow.steps.StepConfigTester;
+import org.jenkinsci.plugins.workflow.structs.DescribableHelper;
+import org.junit.Test;
+import static org.junit.Assert.*;
+import org.junit.Rule;
+import org.jvnet.hudson.test.Issue;
+import org.jvnet.hudson.test.JenkinsRule;
+import org.jvnet.hudson.test.WithoutJenkins;
+
+public class InputStepTest {
+    
+    @Rule public JenkinsRule r = new JenkinsRule();
+
+    @Test public void configRoundTrip() throws Exception {
+        InputStep s1 = new InputStep("hello world");
+        InputStep s2 = new StepConfigTester(r).configRoundTrip(s1);
+        assertEquals(s1.getMessage(), s2.getMessage());
+        assertEquals(s1.getId(), s2.getId());
+        assertEquals(s1.getParameters(), s2.getParameters());
+        assertEquals(s1.getOk(), s2.getOk());
+        assertEquals(s1.getSubmitter(), s2.getSubmitter());
+    }
+
+    @Issue("JENKINS-25779")
+    @WithoutJenkins
+    @Test public void uninstantiate() throws Exception {
+        InputStep s = new InputStep("hello world");
+        assertEquals(Collections.singletonMap("message", s.getMessage()), DescribableHelper.uninstantiate(s));
+    }
+
+}